+++ rwaassetlist.h
@@ -27,6 +27,8 @@
 #include "rwalistview.h"
 #include <QFileInfo>
 #include <QFile>
+#include <QDesktopServices>
+#include <QUrl>
 
 class RwaAssetList : public RwaListView
 {
@@ -43,6 +45,7 @@ public slots:
 protected:
         void keyPressEvent(QKeyEvent *event) override;
         void mousePressEvent(QMouseEvent *event) override;
+        void mouseDoubleClickEvent(QMouseEvent *event) override;
         void mouseReleaseEvent(QMouseEvent *event) override;
         void dragEnterEvent(QDragEnterEvent *event) override;
         void dropEvent(QDropEvent *event) override;
@@ -52,6 +55,7 @@ protected:
 
     private:
         void setCurrentAssetFromCurrentItem();
+        void openAssetExternally(RwaAsset1 *asset);
         void findLastTouchedAsset(RwaAsset1 *asset);
         void setAssetBadges(QListWidgetItem *item, RwaAsset1 *asset);
 signals: